backend/requirements.txt
#Needed pip libraries here
psycopg
flask
python-dotenv
flask_cors